third_party/blink/web_tests/platform/generic/fast/events/hit-test-cache-scrollbar-no-crash-expected.txt

Ensure hit test cache works in correct scenarios of overlay scrollbars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.hitTestCacheHits(document) is expectedCount
PASS successfullyParsed is true

TEST COMPLETE
